cow_backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
cachetools==5.5.2
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
Django==3.2.17
django-admin-interface==0.24.1
django-admin-rangefilter==0.11.1
django-colorfield==0.10.1
django-cors-headers==4.2.0
django-location-field==2.7.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.4
joblib==1.4.2
numpy==2.2.4
pandas==2.2.3
Pillow==10.0.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
scikit-learn==1.4.2
scipy==1.15.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
